requirements.txt
django
psycopg2-binary
python-decouple
djangorestframework
pillow
dj-rest-auth
